Extract URI handling to a URI provider
