+++ core/src/main/java/org/cru/contentscoring/core/servlets/ResourceUrlMapperServlet.java
@@ -4,7 +4,6 @@ import java.io.IOException;
 import java.util.Arrays;
 import java.util.HashSet;
 import java.util.List;
-import java.util.Objects;
 import java.util.Set;
 import java.util.stream.Collectors;
 
@@ -15,9 +14,13 @@ import org.apache.sling.api.SlingHttpServletResponse;
 import org.apache.sling.api.request.RequestParameter;
 import org.apache.sling.api.resource.Resource;
 import org.apache.sling.api.resource.ResourceResolver;
-import org.apache.sling.api.resource.ValueMap;
+import org.apache.sling.api.resource.external.URIProvider;
+import org.apache.sling.api.resource.external.URIProvider.Scope;
+import org.apache.sling.api.resource.external.URIProvider.Operation;
 import org.apache.sling.api.servlets.HttpConstants;
 import org.apache.sling.api.servlets.SlingSafeMethodsServlet;
+import org.cru.contentscoring.core.provider.ContentScoringUriProvider;
+import org.osgi.service.component.annotations.Activate;
 import org.osgi.service.component.annotations.Component;
 
 import com.day.cq.commons.Externalizer;
@@ -31,6 +34,15 @@ import com.fasterxml.jackson.databind.ObjectMapper;
         "sling.servlet.methods=" + HttpConstants.METHOD_GET,
         "sling.servlet.paths=/bin/cru/url/mapper" })
 public class ResourceUrlMapperServlet extends SlingSafeMethodsServlet {
+    private URIProvider uriProvider;
+
+    @Activate
+    public void activate() {
+        if (uriProvider == null) {
+            uriProvider = new ContentScoringUriProvider();
+        }
+    }
+
     @Override
     protected void doGet(final SlingHttpServletRequest request, final SlingHttpServletResponse response) throws IOException {
         RequestParameter[] pathParameters = request.getRequestParameters("path");
@@ -96,56 +108,12 @@ public class ResourceUrlMapperServlet extends SlingSafeMethodsServlet {
 
         Set<String> urls = new HashSet<>();
         for (String path : paths) {
-            urls.add(getUrlFromSlingMap(path, resourceResolver));
+            Resource resource = resourceResolver.getResource(path);
+            if (resource != null) {
+                urls.add(uriProvider.toURI(resource, Scope.EXTERNAL, Operation.READ).toString());
+            }
         }
 
         return urls;
-    }
-
-    private String getUrlFromSlingMap(final String path, final ResourceResolver resourceResolver) {
-        Resource slingMap = determineSlingMap(path, resourceResolver);
-
-        if (slingMap != null) {
-            String protocol = Objects.requireNonNull(slingMap.getParent()).getName();
-            String domain = slingMap.getName();
-            return protocol + "://" + domain + path + ".html";
-        }
-
-        return null;
-    }
-
-    private Resource determineSlingMap(final String path, final ResourceResolver resourceResolver) {
-        String httpsPath = "/etc/map.publish.prod/https";
-        Resource httpsRoot = resourceResolver.getResource(httpsPath);
-
-        Resource slingMap = null;
-        if (httpsRoot != null) {
-            slingMap = loopThroughSlingMaps(httpsRoot, path);
-        }
-
-        String httpPath = "/etc/map.publish.prod/http";
-        Resource httpRoot = resourceResolver.getResource(httpPath);
-        if (httpRoot != null && slingMap == null) {
-            slingMap = loopThroughSlingMaps(httpRoot, path);
-        }
-
-        return slingMap;
-    }
-
-    private Resource loopThroughSlingMaps(final Resource parent, final String path) {
-        if (parent.hasChildren()) {
-            for (Resource child : parent.getChildren()) {
-                if (child.getName().contains("_")) {
-                    continue;
-                }
-
-                ValueMap properties = child.getValueMap();
-                String internalRedirect = properties.get("sling:internalRedirect", String.class);
-                if (internalRedirect != null && path.startsWith(internalRedirect)) {
-                    return child;
-                }
-            }
-        }
-        return null;
     }
 }
